# Pass pointers to C as unsafe.Pointer, not uintptr (memory corruption)