--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -42,7 +42,7 @@ jobs:
       - run: 'go build'
       - run: 'PAYLOAD=`pwd`/mr GODEBUG=efence=1 go test -v'
   "golang-i386":
-    resource_class: large
+    resource_class: xlarge
     docker:
       - image: 32bit/ubuntu:16.04
     steps: