@evanj 2021-02-07
Based on this test, we should probably change the CircleCI tests to run the tests with `GODEBUG=efence=1` to check against these types of extremely subtle issues. I can make an attempt to do that change at some point later.