@Viq111 2021-02-08
Thanks for the PR! 
The `golang-i386` test fail seems unrelated and realted to centos6 EOL instead (Nov 2020). I'll fix it in a separate PR.
For trying to reproduce in CircleCI, I didn't have much success yet. It looks like `efence=1` uses a lot of memory so we wouldn't be able to add it to all tests but we can rerun the tests with that option on but a smaller payload. However pulling your tests (and not the code change) doesn't reproduce the bug in CircleCI: https://app.circleci.com/pipelines/github/DataDog/zstd/56/workflows/2741378d-820d-4224-897e-209f301ac000/jobs/198