@Viq111 2021-02-08
Thanks for the PR!
I tried to import the tests onto the `1.x` (without the fix) and run with `efence` on CircleCI but unfortunatly it still looks like it's hard to replicate (i.e: the tests pass without the fix: https://app.circleci.com/pipelines/github/DataDog/zstd/61/workflows/a2a5fbd9-d76e-4de8-b981-5510372bf198/jobs/213)