@evanj 2021-02-09
Rats you are right. I swear this was crashing reliably for me! Let me investigate a bit more...